# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Job Application Tracker API",
        "version": "1.0.0",
        "description": "REST API for managing companies, job applications, contacts, and interviews."
    },
    "servers": [
        {
            "url": "/",
            "description": "Current server"
        }
    ],
...
    "paths": {
        "/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get all companies",
                "operationId": "getAllCompanies",
                "responses": {
                    "200": {
                        "description": "Companies retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Company"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Create a company",
                "operationId": "createCompany",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Company created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            }
        },
...
        "/applications": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Get all job applications",
                "operationId": "getAllApplications",
                "responses": {
                    "200": {
                        "description": "Applications retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Application"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            },
            "post": {
                "tags": [
                    "Applications"
                ],
                "summary": "Create a job application",
                "operationId": "createApplication",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApplicationInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Application created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            }
        },
...
        "/contacts": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Get all contacts",
                "operationId": "getAllContacts",
                "responses": {
                    "200": {
                        "description": "Contacts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contact"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            },
            "post": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Create a contact",
                "operationId": "createContact",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contact created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            }
        },
...
        "/interviews": {
            "get": {
                "tags": [
                    "Interviews"
                ],
                "summary": "Get all interviews",
                "operationId": "getAllInterviews",
                "responses": {
                    "200": {
                        "description": "Interviews retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Interview"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            },
            "post": {
                "tags": [
                    "Interviews"
                ],
                "summary": "Create a interview",
                "operationId": "createInterview",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InterviewInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Interview created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Interview"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            }
        },
...
        "/auth/github": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign in with GitHub",
                "description": "Open /auth/github in a browser tab to complete login. Swagger Execute cannot complete the interactive redirect flow.",
                "operationId": "loginWithGitHub",
                "responses": {
                    "302": {
                        "description": "Redirect to GitHub for authorization"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            }
        },
        "/auth/status": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Check current login session",
                "operationId": "getAuthenticationStatus",
                "responses": {
                    "200": {
                        "description": "Session status; user is returned when authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthStatus"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "responses": {
            "BadRequest": {
                "description": "Invalid ID or request data",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
...
            "ServerError": {
                "description": "Internal server error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "An authenticated GitHub session is required",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/AuthenticationError"
                        }
                    }
                }
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Unable to process the request"
                    }
                }
            },
            "CompanyInput": {
                "type": "object",
                "required": [
                    "name",
                    "website",
                    "industry",
                    "location",
                    "contactName",
                    "contactEmail"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Acme Technology"
                    },
                    "website": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://www.example.com"
                    },
                    "industry": {
                        "type": "string",
                        "example": "Software"
                    },
                    "location": {
                        "type": "string",
                        "example": "Boise, Idaho"
                    },
                    "contactName": {
                        "type": "string",
                        "example": "Jordan Smith"
                    },
                    "contactEmail": {
                        "type": "string",
                        "format": "email",
                        "example": "jordan@example.com"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Met at a university career fair."
                    }
                }
            },
            "Company": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/CompanyInput"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "_id": {
                                "type": "string",
                                "example": "68cb1234567890abcdef1234"
                            },
                            "createdAt": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updatedAt": {
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    }
                ]
            },
            "ApplicationInput": {
                "type": "object",
                "required": [
                    "position",
                    "companyId",
                    "appliedDate",
                    "status",
                    "workMode",
                    "location",
                    "salaryRange",
                    "jobUrl",
                    "description",
                    "technologies"
                ],
                "properties": {
                    "position": {
                        "type": "string",
                        "example": "Junior Web Developer"
                    },
                    "companyId": {
                        "type": "string",
                        "example": "68cb1234567890abcdef1234"
                    },
                    "appliedDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-09-16"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "saved",
                            "applied",
                            "interviewing",
                            "offer",
                            "rejected",
                            "withdrawn"
                        ],
                        "example": "applied"
                    },
                    "workMode": {
                        "type": "string",
                        "enum": [
                            "remote",
                            "hybrid",
                            "onsite"
                        ],
                        "example": "hybrid"
                    },
                    "location": {
                        "type": "string",
                        "example": "Boise, Idaho"
                    },
                    "salaryRange": {
                        "type": "string",
                        "example": "$60,000-$75,000"
                    },
                    "jobUrl": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://www.example.com/jobs/123"
                    },
                    "description": {
                        "type": "string",
                        "example": "Build and maintain web applications."
                    },
                    "technologies": {
                        "type": "array",
                        "minItems": 1,
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "JavaScript",
                            "Node.js",
                            "MongoDB"
                        ]
                    },
                    "notes": {
                        "type": "string",
                        "example": "Follow up with the recruiter next week."
                    }
                }
            },
            "Application": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ApplicationInput"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "_id": {
                                "type": "string",
                                "example": "68cb1234567890abcdef5678"
                            },
                            "createdAt": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updatedAt": {
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    }
                ]
            },
            "ContactInput": {
                "type": "object",
                "required": [
                    "name",
                    "email",
                    "phone",
                    "role",
                    "companyId"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Jordan Smith",
                        "minLength": 1
                    },
                    "email": {
                        "type": "string",
                        "example": "jordan@example.com",
                        "minLength": 1,
                        "format": "email"
                    },
                    "phone": {
                        "type": "string",
                        "example": "208-555-0100",
                        "minLength": 1
                    },
                    "role": {
                        "type": "string",
                        "example": "Recruiter",
                        "minLength": 1
                    },
                    "companyId": {
                        "type": "string",
                        "example": "6aab43653f32b48f16367f8d",
                        "minLength": 1,
                        "pattern": "^[a-fA-F0-9]{24}$",
                        "description": "ID of an existing company"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Test contact for Acme"
                    }
                }
            },
            "Contact": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ContactInput"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "_id": {
                                "type": "string",
                                "example": "68cb1234567890abcdef1234"
                            },
                            "createdAt": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updatedAt": {
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    }
                ]
            },
            "InterviewInput": {
                "type": "object",
                "required": [
                    "applicationId",
                    "scheduledAt",
                    "type",
                    "interviewer",
                    "location",
                    "status"
                ],
                "properties": {
                    "applicationId": {
                        "type": "string",
                        "example": "6aab443d3f32b48f16367f8e",
                        "minLength": 1,
                        "pattern": "^[a-fA-F0-9]{24}$",
                        "description": "ID of an existing job application"
                    },
                    "scheduledAt": {
                        "type": "string",
                        "example": "2026-10-20T15:00:00Z",
                        "minLength": 1,
                        "format": "date-time",
                        "description": "ISO date-time including seconds and timezone (Z or an offset).",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d{1,3})?(?:Z|[+-]\\d{2}:\\d{2})$"
                    },
                    "type": {
                        "type": "string",
                        "example": "video",
                        "minLength": 1,
                        "enum": [
                            "phone",
                            "video",
                            "onsite"
                        ]
                    },
                    "interviewer": {
                        "type": "string",
                        "example": "Jordan Smith",
                        "minLength": 1
                    },
                    "location": {
                        "type": "string",
                        "example": "https://example.com/meeting",
                        "minLength": 1
                    },
                    "status": {
                        "type": "string",
                        "example": "scheduled",
                        "minLength": 1,
                        "enum": [
                            "scheduled",
                            "completed",
                            "cancelled"
                        ]
                    },
                    "notes": {
                        "type": "string",
                        "example": "Test interview for Junior Web Developer"
                    }
                }
            },
            "Interview": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/InterviewInput"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "_id": {
                                "type": "string",
                                "example": "68cb1234567890abcdef1234"
                            },
                            "createdAt": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updatedAt": {
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    }
                ]
            },
            "AuthenticationError": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Authentication required"
                    },
                    "login": {
                        "type": "string",
                        "example": "/auth/github"
                    }
                }
            },
            "AuthStatus": {
                "type": "object",
                "properties": {
                    "authenticated": {
                        "type": "boolean"
                    },
                    "login": {
                        "type": "string",
                        "example": "/auth/github"
                    },
                    "user": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "username": {
                                "type": "string"
                            },
                            "displayName": {
                                "type": "string"
                            },
                            "profileUrl": {
                                "type": "string"
                            },
                            "avatarUrl": {
                                "type": "string",
                                "nullable": true
                            }
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "sessionCookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "job-tracker.sid",
                "description": "Sign in through /auth/github in this browser. The session cookie is sent automatically to same-origin requests."
            }
        }
    }
}